AVRO-3369 Move default reader to own file
